% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cats_function.R
\name{cats_function}
\alias{cats_function}
\title{A Cat Function}
\usage{
cats_function(feed = TRUE)
}
\arguments{
\item{feed}{Have you fed cats? Defaults to TRUE.}
}
\description{
This function reminds you to feed your cats before going to bed.
}
\examples{
cats_function()
}
\keyword{cats}
